Made small changes to be able to run on Python3
PyMad behaves weird in Python3, whatever I tried, I'd get '-1' as the duration, so I used tinytag instead which supports audio files other than mp3 which is something we definitely want anyways. Other small fixes include updating ffmpeg to output to full HD, and
ditch .avi as it's not 2000 anymore, it now uses a very standard and default x264 output with mp4 as the container.